+++ src/routes/DashboardRoutes.tsx
@@ -11,6 +11,7 @@ import {
   TeamDetails,
   FleetMap,
   Organization,
+  OrganizationMembers,
   OrganizationIntegrations,
   PMTemplates,
   PMTemplateEditor,
@@ -45,6 +46,7 @@ export const dashboardRouteElements = (
     <Route path="/teams/:teamId" element={<TeamDetails />} />
     <Route path="/fleet-map" element={<FleetMap />} />
     <Route path="/organization" element={<Organization />} />
+    <Route path="/organization/members" element={<OrganizationMembers />} />
     <Route path="/organization/integrations" element={<OrganizationIntegrations />} />
     <Route path="/pm-templates" element={<PMTemplates />} />
     <Route path="/pm-templates/new" element={<PMTemplateEditor />} />